refactor: 파일 상수 패키지 분리
