+++ src/main/java/store/config/ProductsFileInitializer.java
@@ -1,9 +1,9 @@
 package store.config;
 
-import static store.constants.FileConstants.NON_PROMOTION;
-import static store.constants.FileConstants.PRODUCT_FILE_SIZE;
-import static store.constants.FileConstants.SPLIT_REGEX;
-import static store.constants.FileErrorMessage.INVALID_FILE_CONTENT_FORM;
+import static store.constants.fileConstants.FileConstants.NON_PROMOTION;
+import static store.constants.fileConstants.FileConstants.PRODUCT_FILE_SIZE;
+import static store.constants.fileConstants.FileConstants.SPLIT_REGEX;
+import static store.constants.fileConstants.FileErrorMessage.INVALID_FILE_CONTENT_FORM;
 
 import java.util.Arrays;
 import java.util.List;
